chore: Contexts organized
