+++ pages/agreements/index.js
@@ -1,5 +1,5 @@
 import UserCheck from "../../components/UserCheck/UserCheck"
-import { AgreementContractProvider } from "../../components/ContractContext";
+import { AgreementContractProvider } from "../../contexts/ContractContext";
 
 export default function Agreements() {
   return (
